Don't use the stack after freeing it.
When the stack size is >=128k, malloc mmaps the stack. This means that
once the stack is freed, any access to it will become a segfault, so the
trick of freeing the stack, and then using the space for a little bit
longer since we know nothing else could take the space won't work
anymore.

Therefore, instead of immediately deleting the coroutine, we add an idle
watcher to the event loop so that the next time there are spare cycles
on the system, we clean up all the dead coroutines. This could
potentially lead to DoS issues on a busy server, so it might need to be
changed so that dead coroutines are cleaned up in an ev_check watcher
(either one watcher per coro or one watcher that checks a list every
iteration).